+++ pierky/arouteserver/peering_db.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2017-2019 Pier Carlo Chiodi
+# Copyright (C) 2017-2020 Pier Carlo Chiodi
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -187,6 +187,33 @@ class PeeringDBNetIXLan(PeeringDBInfo):
 
     def _get_peeringdb_url(self):
         return self.PEERINGDB_URL.format(ixlanid=self.ixlanid)
+
+class PeeringDBNetNeverViaRouteServers(PeeringDBInfo):
+
+    PEERINGDB_URL = "https://www.peeringdb.com/api/net?info_never_via_route_servers=1"
+
+    EXPIRY_TIME_TAG = "pdb_info"
+
+    def __init__(self, **kwargs):
+        PeeringDBInfo.__init__(self, **kwargs)
+
+        self.networks = []
+
+    def load_data(self):
+        logging.debug("Getting 'never via route-servers' networks from PeeringDB")
+
+        PeeringDBInfo.load_data(self)
+
+        for network in self.raw_data:
+            self.networks.append({
+                "asn": network["asn"]
+            })
+
+    def _get_object_filename(self):
+        return "peeringdb_neverviarouteservers.json"
+
+    def _get_peeringdb_url(self):
+        return self.PEERINGDB_URL.format()
 
 class PeeringDBIXList(PeeringDBInfo):